backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
sentence-transformers
unidecode
nltk
gunicorn
faiss-cpu
tqdm
supabase
python-dotenv
